+++ src/novel/dto/index.ts
@@ -1,4 +1,4 @@
-export { default as CreateNovelDto } from './create-novel.dto';
-export { default as FindNovelListCategoryDto } from './find-novel-list-category.dto';
-export { default as PaginationMetaDto } from './pagination.meta.dto';
-export { default as FindNovelFeedDto } from './find-novel-feed.dto';
+export { default as CreateNovelDto } from './request/create-novel.dto';
+export { default as FindNovelListCategoryDto } from './request/find-novel-list-category.dto';
+export { default as PaginationMetaDto } from './response/pagination.meta.dto';
+export { default as FindNovelFeedDto } from './response/find-novel-feed.dto';
